renaming package substratemanager to subi
