+++ node/client_pool.go
@@ -2,7 +2,7 @@ package client
 
 import (
 	"github.com/pkg/errors"
-	substratemanager "github.com/threefoldtech/grid3-go/substrate_manager"
+	substratemanager "github.com/threefoldtech/grid3-go/subi"
 	"github.com/threefoldtech/zos/pkg/rmb"
 )
