+++ subi/errors.go
@@ -1,4 +1,4 @@
-package substratemanager
+package subi
 
 import (
 	"github.com/pkg/errors"
