+++ subi/substrate_manager.go
@@ -1,4 +1,4 @@
-package substratemanager
+package subi
 
 import (
 	"context"
